Move noEmptyRevokedCertificatesList
