+++ linter/lints/crl/lints.go
@@ -98,19 +98,6 @@ func LintCRL(lintCRL *crl_x509.RevocationList) *zlint.ResultSet {
 	return &rset
 }
 
-// noEmptyRevokedCertificatesList checks RFC 5280, Section 5.1.2.6:
-// When there are no revoked certificates, the revoked certificates list MUST be
-// absent.
-func noEmptyRevokedCertificatesList(crl *crl_x509.RevocationList) *lint.LintResult {
-	if crl.RevokedCertificates != nil && len(crl.RevokedCertificates) == 0 {
-		return &lint.LintResult{
-			Status:  lint.Error,
-			Details: "If the revokedCertificates list is empty, it must not be present",
-		}
-	}
-	return &lint.LintResult{Status: lint.Pass}
-}
-
 // hasAKI checks RFC 5280, Section 5.2.1:
 // Conforming CRL issuers MUST use the key identifier method, and MUST include
 // this extension in all CRLs issued.
